Nauru

Nauru

NRU(level)

Source

https://github.com/covid19datahub/COVID19/blob/master/R/iso_NRU.R

Data Sources

Level 1

Johns Hopkins Center for Systems Science and Engineering: confirmed cases, deaths, recovered.

Our World in Data: tests, total vaccine doses administered, people with at least one vaccine dose, people fully vaccinated, hospitalizations, intensive care.